requirements.txt
requests
streamlit
pandas
numpy
humanize
pillow